requirements.txt
python-dotenv
discord.py
fasteners